scanpath similarity methods
scanpath similarity methods moved to seprate file _scanpathsim.py
