+++ PyEyeSim/__init__.py
@@ -30,6 +30,7 @@ class EyeData:
     from ._dataproc import GetParams,GetStimuli,GetFixationData,GetDurations,GetGroups,GetCats,GetSaccades,SaccadeSel,GetEntropies,InferSize,Heatmap,FixCountCalc
     from ._stats import AngleCalc,AngtoPix,PixdoDeg,Entropy,FixDurProg,BinnedCount,GetInddiff,GetInddiff_v2,RunDiffDivs,GetBinnedStimFixS,StatPDiffInd2,StatPDiffInd1,CalcStatPs,CalcRets,CalcImmRets,BinnedDescriptives
     from ._comparegroups import CompareGroupsFix,CompareGroupsHeatmap,CompareWithinGroupsFix,FixDurProgGroups,BinnedDescriptivesGroups,CompareGroupsMat
+    from ._scanpathsimhelper import AOIFix,SacSimPipeline,SacSim1Group
 
     try: 
     	from ._hmm import DataArrayHmm,MyTrainTest,FitLOOHMM,FitVisHMM,FitVisHMMGroups,HMMSimPipeline
@@ -222,83 +223,6 @@ class EyeData:
   
     
    
-
-
-
-    
-    def AOIFix(self,p,FixTrialX,FixTrialY,nDivH,nDivV,InferS=1):
-        """ given a sequence of X,Y fixation data and AOI divisions, calculate static N and p matrix) """ 
-        nAOI=nDivH*nDivV
-        AOInums=np.arange(nAOI).reshape(nDivV,nDivH)
-        NFix=len(FixTrialX)  # num of data points
-        # set AOI bounds
-       # print(p,SizeGendX)
-        if InferS==0:
-            AOIboundsH=AOIbounds(0, self.x_size,nDivH)       
-            AOIboundsV=AOIbounds(0, self.y_size,nDivV)  
-        else:
-            AOIboundsH=AOIbounds(self.boundsX[p,0], self.boundsX[p,1],nDivH)       
-            AOIboundsV=AOIbounds(self.boundsY[p,0], self.boundsY[p,1],nDivV)   
-       
-        # set parameters & arrays to store data
-        StatPtrial=np.zeros(nAOI) # static probabilities.
-        StatNtrial=np.zeros(nAOI) # static counts.
-    
-       
-        WhichAOIH=np.zeros(NFix)
-        WhichAOIV=np.zeros(NFix)
-        for x in range(NFix):
-            WhichAOIH[x]=CheckCoor(AOIboundsH,FixTrialX[x]) # store which horizontal AOI each fixation is
-            WhichAOIV[x]=CheckCoor(AOIboundsV,FixTrialY[x]) # store which vertical AOI each fixation is
-    
-        WhichAOI=np.zeros(NFix)
-        WhichAOI[:]=np.NAN
-        for x in range(NFix):
-            if WhichAOIV[x]>-1 and WhichAOIH[x]>-1:   # only use valid idx
-                WhichAOI[x]=AOInums[np.intp(WhichAOIV[x]),np.intp(WhichAOIH[x])]  # get combined vertival and horizontal
-        for st in range(nAOI): # gaze transition start
-            StatNtrial[st]=np.sum(WhichAOI==st)  # get count in AOI
-            StatPtrial[st]=np.sum(WhichAOI==st)/np.sum(np.isfinite(WhichAOI)) # calculate stationary P for each AOI    
-        return NFix,StatPtrial,StatNtrial
-    
-    
-         
-       ## function still missing
-
-
-    def SacSimPipeline(self,divs=[4,5,7,9],Thr=5):
-        SaccadeObj=self.GetSaccades()
-        StimSims=np.zeros((len(divs),self.np))
-        StimSimsInd=np.zeros((len(divs),self.ns,self.np))
-
-        for cd,ndiv in enumerate(divs):
-            sacDivSel=self.SaccadeSel(SaccadeObj,ndiv)
-            SimSacP=self.SacSim1Group(sacDivSel,ndiv,Thr=Thr)
-            StimSimsInd[cd,:,:]=np.nanmean(np.nanmean(np.nanmean(SimSacP,4),3),0)
-            StimSims[cd,:]=np.nanmean(np.nanmean(np.nanmean(np.nanmean(SimSacP,4),3),0),0)
-        return StimSims,np.nanmean(StimSimsInd,0)
-   
-
-
-
-
-    def SacSim1Group(self,Saccades,nDiv,Thr=5):
-        ''' calculate saccade similarity for each stimulus, betwween each pair of participants '''
-        nHor,nVer=nDiv,nDiv
-        SimSacP=np.zeros((self.ns,self.ns,self.np,nHor,nVer))  
-        SimSacP[:]=np.NaN
-        for s1 in range(self.ns):
-            for s2 in range(self.ns):
-                if s1!=s2:
-                    for p1 in range(self.np):
-                        if self.nsac[s1,p1]>5 and self.nsac[s2,p1]>5:                    
-                            for h in range(nHor):
-                                for v in range(nVer):
-                                    if len(Saccades[s1,p1,h,v])>0 and len(Saccades[s2,p1,h,v])>0:
-                                            
-                                        simsacn=CalcSim(Saccades[s1,p1,h,v],Saccades[s2,p1,h,v],Thr=Thr)
-                                        SimSacP[s1,s2,p1,h,v]=simsacn/(len(Saccades[s1,p1,h,v])+len(Saccades[s2,p1,h,v]))
-        return SimSacP
